Modularize Soniqo transcription crate
Separate model identity, public payloads, response synthesis, platform backends, and tests behind the existing crate API without changing transcription behavior.
